# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SISAM - Sistema de Avaliação Municipal
  version: 2.0.0
  description: |
    API do sistema educacional SISAM / Gestor Escolar da SEMED São Sebastião da Boa Vista.

    O sistema abrange avaliações municipais (SISAM), gestão escolar completa, portal do professor,
    reconhecimento facial para frequência, site institucional, publicações oficiais e boletim público.

    ## Autenticação
    A autenticação é feita via cookie JWT httpOnly (`token`) definido no login.
    Todas as rotas protegidas requerem o cookie `token` válido.

    ## Tipos de Usuário
    - **administrador** — Acesso total ao sistema
    - **tecnico** — Acesso total (equivalente ao administrador)
    - **polo** — Acesso ao seu polo e escolas vinculadas
    - **escola** — Acesso apenas à sua própria escola
    - **professor** — Acesso ao portal do professor (turmas vinculadas)
    - **editor** — Acesso ao editor de notícias do site
    - **publicador** — Acesso ao módulo de publicações oficiais

    ## Padrões
    - IDs são UUIDs v4
    - Respostas de erro: `{ mensagem: string }`
    - Paginação: `pagina` (default 1), `limite` (default 50, max 200)
    - Ano letivo: string de 4 dígitos (ex: "2026")
  contact:
    name: Junielson Farias
servers:
  - url: /api
    description: API base path
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: JWT token em cookie httpOnly
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
...
paths:
...
  /admin/importar:
    post:
      tags: [Admin/Importações]
      summary: Importar resultados via planilha
      description: Upload de planilha Excel com resultados de avaliação SISAM.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [arquivo]
              properties:
                arquivo:
                  type: string
                  format: binary
                ano_letivo:
                  type: string
                avaliacao_id:
                  $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Importação realizada
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensagem:
                    type: string
                  importacao:
                    type: object
...